+++ .eslintignore
@@ -215,12 +215,11 @@ packages/app-desktop/gui/MenuBar.js
 packages/app-desktop/gui/MultiNoteActions.js
 packages/app-desktop/gui/Navigator.js
 packages/app-desktop/gui/NoteContentPropertiesDialog.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/CodeMirror.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/Editor.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/Toolbar.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/styles/index.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/index.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/types.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useContextMenu.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useCursorUtils.test.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useCursorUtils.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useEditorSearch.js
@@ -232,12 +231,13 @@ packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useLineSorting.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useListIdent.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useScrollHandler.js
 packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useScrollUtils.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/CodeMirror6.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/Editor.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/Toolbar.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/useEditorCommands.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/useScrollHandler.js
-packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror6/useStyles.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useStyles.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/utils/useWebviewIpcMessage.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v5/CodeMirror.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v5/Editor.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v6/CodeMirror.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v6/Editor.js
+packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v6/useEditorCommands.js
 packages/app-desktop/gui/NoteEditor/NoteBody/PlainEditor/PlainEditor.js
 packages/app-desktop/gui/NoteEditor/NoteBody/TinyMCE/TinyMCE.js
 packages/app-desktop/gui/NoteEditor/NoteBody/TinyMCE/styles/index.js
