+++ packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v5/Editor.tsx
@@ -12,15 +12,15 @@ import 'codemirror/addon/scroll/annotatescrollbar';
 import 'codemirror/addon/search/matchesonscrollbar';
 import 'codemirror/addon/search/searchcursor';
 
-import useListIdent from './utils/useListIdent';
-import useScrollUtils from './utils/useScrollUtils';
-import useCursorUtils from './utils/useCursorUtils';
-import useLineSorting from './utils/useLineSorting';
-import useEditorSearch from './utils/useEditorSearch';
-import useJoplinMode from './utils/useJoplinMode';
-import useKeymap from './utils/useKeymap';
-import useExternalPlugins from './utils/useExternalPlugins';
-import useJoplinCommands from './utils/useJoplinCommands';
+import useListIdent from '../utils/useListIdent';
+import useScrollUtils from '../utils/useScrollUtils';
+import useCursorUtils from '../utils/useCursorUtils';
+import useLineSorting from '../utils/useLineSorting';
+import useEditorSearch from '../utils/useEditorSearch';
+import useJoplinMode from '../utils/useJoplinMode';
+import useKeymap from '../utils/useKeymap';
+import useExternalPlugins from '../utils/useExternalPlugins';
+import useJoplinCommands from '../utils/useJoplinCommands';
 
 import 'codemirror/keymap/emacs';
 import 'codemirror/keymap/vim';
