+++ packages/app-desktop/gui/NoteEditor/NoteBody/CodeMirror/v6/Editor.tsx
